Crate pallet_donations

Donations Module

The Donations module allows accounts to make donations (aka tips) to other accounts, spaces, or posts.

Structs

Donation
A struct that describes a single donation made by an account.

DonationById
Donation details by its id (key).

DonationIdsByRecipient
Ids of all donations received by this recipient (key). A recipient can be either account, space or post.

DonationIdsBySupporter
Ids of all donations made by a supporter account (key).

DonationSettings

DonationSettingsByRecipient
A custom donation settings for a certain recipient (key). This means that any account, space or post can set up a custom donation settings that will be applied to future donations to this recipient.å

DonationSettingsUpdate

DonationWalletByRecipient
A custom wallet for a certain recipient (key). This means that any account, space or post can set up a custom wallet address that will be used for future donations to this recipient.

Module

NextDonationId
An id for the next donation.

Enums

Call
Dispatchable calls.

DonationRecipient

Error

RawEvent
Events for this module.

Traits

Trait

Type Definitions

DonationId

Event
RawEvent specialized for the configuration Trait